+++ ethclient/ethclient_test.go
@@ -19,15 +19,12 @@ package ethclient
 import (
 	"bytes"
 	"context"
-	"crypto/ecdsa"
 	"errors"
 	"fmt"
 	"math/big"
 	"reflect"
 	"testing"
 	"time"
-
-	"github.com/stretchr/testify/require"
 
 	ethereum "github.com/autonity/autonity"
 	"github.com/autonity/autonity/common"
@@ -39,7 +36,6 @@ import (
 	"github.com/autonity/autonity/eth"
 	"github.com/autonity/autonity/eth/ethconfig"
 	"github.com/autonity/autonity/node"
-	"github.com/autonity/autonity/p2p/enode"
 	"github.com/autonity/autonity/params"
 	"github.com/autonity/autonity/rpc"
 )
@@ -58,75 +54,6 @@ var (
 	// _ = ethereum.PendingStateEventer(&Client{})
 	_ = ethereum.PendingContractCaller(&Client{})
 )
-
-func bfs(
-	source *enode.Node, dis map[*enode.Node]int, localNodes map[*enode.Node]*enode.LocalNode,
-	nodes []*enode.Node, adjacentNodes func([]*enode.Node, *enode.LocalNode) []*enode.Node,
-) {
-	for key := range dis {
-		dis[key] = -1
-	}
-	// enque source
-	dis[source] = 0
-	queue := make([]*enode.Node, 0, len(nodes))
-	queue = append(queue, source)
-	for len(queue) > 0 {
-		// pop
-		node := queue[0]
-		queue = queue[1:]
-		connections := adjacentNodes(nodes, localNodes[node])
-		for _, peer := range connections {
-			if d, ok := dis[peer]; !ok || d < 0 {
-				// enque adjacent nodes
-				dis[peer] = dis[node] + 1
-				queue = append(queue, peer)
-			}
-		}
-	}
-}
-
-func TestEthExecutionLayerGraph(t *testing.T) {
-	ethServer := &eth.Ethereum{}
-	db, err := enode.OpenDB("")
-	require.NoError(t, err)
-	nodeCount := int(max(100, params.TestAutonityContractConfig.MaxCommitteeSize))
-	nodes := make([]*enode.Node, 0, nodeCount)
-	localNodes := make(map[*enode.Node]*enode.LocalNode)
-	privateKeys := make(map[*ecdsa.PrivateKey]bool)
-	targetDiameter := 2
-	for n := 0; n < nodeCount; n++ {
-		for {
-			privateKey, err := crypto.GenerateKey()
-			require.NoError(t, err)
-			if _, ok := privateKeys[privateKey]; !ok {
-				newEnode := "enode://" + string(crypto.PubECDSAToHex(&privateKey.PublicKey)[2:]) + "@3.209.45.79:30303"
-				newNode, err := enode.ParseV4(newEnode)
-				require.NoError(t, err)
-				require.NotEqual(t, nil, newNode)
-				nodes = append(nodes, newNode)
-				// related localNode
-				// db is not used here, so a single db for all nodes
-				localNode := enode.NewLocalNode(db, privateKey, nil)
-				localNodes[newNode] = localNode
-				privateKeys[privateKey] = true
-				break
-			}
-		}
-		for _, node := range nodes {
-			// check if max distance from node to any other node in the graph is targetDiameter
-			dis := make(map[*enode.Node]int)
-			bfs(node, dis, localNodes, nodes, ethServer.RequestSubset)
-			for _, peer := range nodes {
-				if d, ok := dis[peer]; !ok || d < 0 {
-					t.Fatalf("Graph with %v nodes not connected", n)
-				}
-				if dis[peer] > targetDiameter {
-					t.Fatalf("Graph with %v nodes has diameter more than %v", n, targetDiameter)
-				}
-			}
-		}
-	}
-}
 
 func TestToFilterArg(t *testing.T) {
 	blockHashErr := fmt.Errorf("cannot specify both BlockHash and FromBlock/ToBlock")
